requirements.txt
Fabric==1.6.0
Flask==0.9
Jinja2==2.6
Pygments==1.6
SQLAlchemy==0.8.1
WTForms==1.0.4
Werkzeug==0.8.3
bpython==0.12
cuisine==0.6.3
dogstatsd-python==0.2.1
flake8==2.0
mccabe==0.2.1
paramiko==1.10.1
pep8==1.4.5
postmonkey==1.0a4
psycopg2==2.5
pycrypto==2.6
pyflakes==0.7.2
requests==1.2.0
ssh==1.8.0
uWSGI==1.9.10
wsgiref==0.1.2
